karoshi_servers_add_role_fm.cgi - reverse proxy server module logic
This improves the logic for the reverse proxy server module which previously had to be on its own server.
Now you can either apply the reverse proxy server module or any of the modules that use apache.
Additionally the reverse proxy module now copies in any existing reverse proxy rules.