Remove setting home directory path for users in ldap
--home-directory='\\'$SERVER'\'$NEWUSERNAME

Setting the home directory path in ldap was added to allow mac clients
to automatically map user home areas. This caused problems on windows
clients setting their temp folders to a network share which would
automatically unmount after a period of inactivity. Although the share
would automatically remount the windows temp folder could still cause
problems.

Now we have a better way to map network shares on mac clients we no
longer need to set the home directory path as this is done as part of a
mac logon script.