Download gitlab archive from LFS
The setup for gitlab now checks that we have a valid archive in case it
is a placeholder for LFS and downloads it if needed. This allows the
module to be rolled out in a patch to existing servers. Other modules
could use the same method for saving space on the Karoshi Server iso
image.