Download Gitlab-CE package from package host
The Gitlab-CE 8.1.4 (Omnibus, Ubuntu Trusty amd64) will now be downloaded from the host provided at the Gitlab website (https://packages.gitlab.com/gitlab/gitlab-ce). The existing file will be checked against an md5 checksum to test it's validity, if the package is invalid it will be (re)downloaded. The host is currently signed using HTTPS however if the package is downloaded, it will still be compared against an md5 checksum to check if the file is correct.