add_user sets home directory attribute
The add_user script again sets the home-directory attribute in Active
Directory. This was originally added in for macs to map home drives but
taken out after full mac logon support was added. It seems that this
parameter is also required in windows for folder redirection.

The parareter below has been added to the samba-tool command:
--home-directory='\\'$SERVER'\'$NEWUSERNAME