Isolinux boot splash fix and add_user home customisation
Isolinux boot splash resolution is now set to get the boot splash to
work correctly.

add_user_fm.cgi has been changed so that the username styles are now
restricted for business and home server installs. This is due to what
seemed like duplicate username styles for non student users where the
year of entry is no longer used.

jsmith09 - 09jsmith : Fine for student accounts but does not make sense
when the year of entry is not used in the username style.