Gluster volume boot checks
This fixes a problem with servers using gluster where one of the servers
is offline and the second server is rebooted. In this instance the
gluster volume would not start because the second gluster server cannot
be contacted and as a result any data on the gluster volume would not be
available.

The mount_gluster_volumes script checks any gluster volumes listed in
fstab to see if they are mounted at the end of the boot process. Any
stopped gluster volumes are started with the force option and then
mounted.

This is primarily for high availability on domain controllers to ensure
that essential samba shares are online in the event of only domain
controller being online and is restarted for any reason.

This also fixes the problem of starting servers after a shutdown of all
servers .One of the servers will be up first and will not start the
gluster volume since it cannot contact the other gluster
server(s) since they have not booted yet.