Fix for excessive inodes with internet logging
The logging system was using one file per user and per site which was
adding up to a large number of files which on a small partition would
result in all of the inodes being used up.

Previous days logs are now stored as one tar archive per section instead
of individual tar files.