Sysvol permissions, password expiry
Sysvol is now watched using inotify to ensure that the permissions are
correct when files are edited from a windows client via gpmc which sets
acls on the files which are not neccesarily correct - particulary when
using additional domain controllers which may have different uid/ gids
for the windows accounts which are non posix accounts.

Password expiry added to the default user settings page and to bulk user
actions.