force_reboot
Script added to be used in circumstances where a normal reboot often
causes the server to hang. This script does its best to shut down
services and unmount devices and the forcibly reboots.

At the moment this script is only called by the update_server script and
is only used if the force_reboot flag file exists in /opt/karoshi/

updateserver

Modified to check for the force_reboot flag file and use the
force_reboot script if it exists when the server is required to restart.

karoshi_servers_remove_role

Bug fix to remove network shares when a server is removed.

ups_status
Modified to use short names for servers in tool tips.