Fix for backup assign and dhcp reservations
backup_assign - now schedules home areas if they are glustered.
Previously glustered home areas were left out of the backup.

dhcp_bans and dhcp_reservations modified to fix a problem with spaces in
client names which were being converted to a + and are not allowed in
the dhcp conf file resulting in the dhcp service not restarting.