Update Servers page modified with a force reboot option
Selecting force reboot creates a flag file in /opt/karoshi/force_reboot.
This flag file is detected when upgrades are carried out and in the
event of the server needing to be rebooted the force_reboot script will
be run instead of using the shutdown command.

The force_reboot script found in
/opt/karoshi/serversetup/distro/ubuntu/scripts/force_reboot shuts down
all required services and unmounts all storage. It then uses reboot -f
for the reboot.