Fix for joining additional domain controllers
This is a work around so that when additional domain controllers are
added the correct dns entries are also added.

This work around temporarily converts the dns on the main server to use
internal samba4 dns. After the additional server is added the main
server is converted back to use bind.