sysvsupport modified to also force systemd to use sysv for some services
This change is due to upstart being removed in the next lts release. All
services used by Karoshi will still be controlled by sysv scripts which
means that the reserveports feature will still work and controls that
are in place to get the timings of starting services with ZFS and
glusterfs will still be correct.