Windows users now default to a roaming profile
New users are now created with a roaming profile by default on new
Karoshi server 11.1 installs.

The windows client roaming profile page now creates or deleles a flag
file as needed so that any new users created will have roaming profiles
if this is the default option for the group.