Bulk User Creation modified so the CSV file can have column headers
The CSV file can be the same format as before or it can have the first
row as headers allowing the columns to be in any order.

The posible headers on the first row are:

forename, surname, enrolment-number, username, primary-group,
secondary-groups, change-password-on-logon,password

The forname and surname headers must be used otherwise the bulk user
creation will default to not using the first row as headers.